requirements.txt
beautifulsoup4==4.13.3
duckduckgo_search==2025.4.4
faster_whisper==1.1.1
kokoro==0.9.4
loguru==0.7.3
numpy==2.2.4
ollama==0.4.7
pytest==8.3.5
PyYAML==6.0.2
trafilatura==2.0.0
transformers>=4.50.2
tokenizers>=0.21.0
huggingface-hub>=0.29.3
numpy>=2.2.4
soundfile>=0.13.1
python-dotenv>=1.1.0       # Environment variables
asyncio>=3.4.3             # Async support
pytest-asyncio>=0.26.0     # Async testing support
webrtcvad>=2.0.10         # Voice activity detection
misaki[en,zh]>=0.9.3   # G2P for multiple languages support
#from misaki import en, zh, espeak
#fallback = espeak.EspeakFallback(lang="ja")
# Use fallback for Japanese text
phonemizer-fork

nltk #memory

# Added dependencies for enhanced embedding system
sentence-transformers>=2.2.2  # Efficient sentence embeddings
pandas              # Data manipulation and analysis
torch>=2.0.0                 # PyTorch for embedding models
scikit-learn>=1.2.0         # For vector operations and similarity metrics
faiss-cpu>=1.7.4            # Fast similarity search (optional, for larger memory stores)
chromadb           # Vector database for embeddings
matplotlib           # For visualization of embeddings
tabulatura         # For tabular data handling